Refactors with main injecting implementation details on App start
