+++ sources/encryptionService.go
@@ -1,0 +1,5 @@
+package main
+
+type EncryptionService interface {
+	comparePasswords(ciphered string, plain string) bool
+}
