+++ sources/main.go
@@ -1,0 +1,10 @@
+package main
+
+func main() {
+	app := App{ 
+				database: NewPostGreSQLDatabase(),
+				router: NewMuxRouter(),
+				encryptionService: NewBCryptEncryptionService(),
+			}
+	app.run()
+}
