+++ sources/router.go
@@ -1,0 +1,10 @@
+package main
+
+import "net/http"
+
+type Router interface {
+	registerRoutes(routes []Route)
+	start() error
+	userIDForRequest(r *http.Request) (int, error)
+	itemIDForRequest(r *http.Request) (int, error) 
+}
